Refactor to remove sqlite dependency
