+++ gtfs-realtime.gemspec
@@ -23,11 +23,11 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency "bundler", "~> 1.13"
   spec.add_development_dependency "rake", "~> 10.0"
+  spec.add_development_dependency "sqlite3"
   spec.add_development_dependency "rspec", "~> 3.0"
   spec.add_development_dependency "webmock"
 
   spec.add_dependency "gtfs-realtime-bindings"
   spec.add_dependency "gtfs"
   spec.add_dependency "sequel"
-  spec.add_dependency "sqlite3"
 end
