+++ lib/gtfs/realtime/configuration.rb
@@ -6,8 +6,12 @@ module GTFS
       def database_path=(new_path)
         @database_path = new_path
 
-        # reinitialize the database with the new path
+        # now that we know the DB path, we can initialize the database
+        require 'gtfs/realtime/database'
         GTFS::Realtime::Database.path = database_path
+
+        # now that we have a database, initialize all the other models
+        require 'gtfs/realtime/bootstrap'
       end
     end
   end
